requirements.txt
streamlit==1.31.0
langchain==0.1.11
langchain-community==0.0.27
python-dotenv==1.0.1
ibm-watsonx-ai>=1.3.3
pymupdf==1.23.26
sentence-transformers>=2.2.2
faiss-cpu==1.7.4
pdf2image==1.17.0
pillow==10.3.0